pyproject.toml
[project]
name = "web-scraper"
version = "0.1.0"
description = "Simple web scraper that converts URLs to markdown"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "httpx>=0.25.0",
    "readability-lxml>=0.8.1",
    "markdownify>=0.11.6",
    "python-multipart>=0.0.6",
    "uvicorn>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

